Fix: remove files and put everything on main.go
